+++ filename.js
@@ -1,5 +1,0 @@
-var lowercase = require('./lowercase');
-var uppercase = require('./uppercase');
-var numbers = require('./numbers');
-
-module.exports = '_-,' + numbers + lowercase + uppercase;